Turn off LED effects with Fn + led
I found that I sometimes want to turn off LED effects. It is cumbersome
to do this by cycling all LED patterns, especially since there are some
that you can only see when pressing keys.

Apparently Fn + led is mapped to Caps Lock. I don't use that and have
turned this off in my OS already. Also, I think that if you want to put
this on the keyboard somewhere it might be better off on Fn + Shift or
so.